pyproject.toml
[tool.poetry]
name = "src"
version = "0.0.1"
description = "Not really a project. Just some S3 scripts."
authors = ["Thomas Fraunholz <tf@sourcemaker.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0.0"
boto3 = "^1.34.80"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = "^22.10.0"
isort = "^5.10.1"
coverage = "^6.5.0"
pre-commit = "^2.20.0"
mypy = "^0.991"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# https://github.com/psf/black
line-length = 79
target_version = ['py310']
exclude = '''
(
  /(
    \ .git
    | .github
    | \.pytest_cache
    | htmlcov
    | venv
    | .venv
  )/
)
'''

[tool.isort]
# https://github.com/timothycrosley/isort
# https://github.com/timothycrosley/isort/wiki/isort-Settings
line_length = 79
indent = '    '
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0

[tool.coverage]
# https://github.com/nedbat/coveragepy

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "# noqa",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]
